# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ENACT Protocol Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Current state of the ENACT job
 */
export type JobState = (typeof JobState)[keyof typeof JobState];

export const JobState = {
  OPEN: "OPEN",
  FUNDED: "FUNDED",
  SUBMITTED: "SUBMITTED",
  COMPLETED: "COMPLETED",
  DISPUTED: "DISPUTED",
  CANCELLED: "CANCELLED",
} as const;

export type JobType = (typeof JobType)[keyof typeof JobType];

export const JobType = {
  TON: "TON",
  USDT: "USDT",
} as const;

export interface Job {
  /** TON contract address */
  address: string;
  /** Sequential job ID from factory */
  jobId: number;
  state: JobState;
  type: JobType;
  /** Client wallet address */
  client: string;
  /** Provider wallet address (null if not taken) */
  provider?: string | null;
  /** Evaluator wallet address */
  evaluator: string;
  /** Budget in TON (string to avoid float precision issues) */
  budgetTon: string;
  /** Budget in USDT (for Jetton jobs) */
  budgetUsdt?: string | null;
  /** IPFS/TON Storage hash of job description */
  descriptionHash: string;
  /** Human-readable description (if available) */
  descriptionText?: string | null;
  /** IPFS/TON Storage hash of submitted result */
  resultHash?: string | null;
  /** Result type code */
  resultType?: number | null;
  /** Job timeout in seconds */
  timeout: number;
  /** Evaluation timeout in seconds */
  evalTimeout: number;
  /** Unix timestamp of creation */
  createdAt: number;
  /** Unix timestamp when result was submitted */
  submittedAt?: number | null;
  /** Link to TON explorer for this contract */
  tonviewerUrl: string;
}

export interface JobListResponse {
  jobs: Job[];
  total: number;
  offset: number;
  limit: number;
}

export type ProtocolStatsByState = {
  OPEN: number;
  FUNDED: number;
  SUBMITTED: number;
  COMPLETED: number;
  DISPUTED: number;
  CANCELLED: number;
};

export interface ProtocolStats {
  totalJobs: number;
  totalTonJobs: number;
  totalUsdtJobs: number;
  byState: ProtocolStatsByState;
  /** Total volume in TON */
  totalVolumeTon: string;
  /** Percentage of completed jobs (0-100) */
  completionRate: number;
  /** Average evaluation time in hours */
  averageEvalTimeHours?: number | null;
}

export interface CreateJobRequest {
  /** Evaluator wallet address */
  evaluator: string;
  /** Budget in TON */
  budgetTon: string;
  /** Job description text */
  description: string;
  /** Job timeout in hours */
  timeoutHours?: number;
  /** Evaluation timeout in hours */
  evalTimeoutHours?: number;
  jobType?: JobType;
}

export interface CreateJobLinkResponse {
  /** Tonkeeper deeplink for transaction signing */
  tonkeeperLink: string;
  /** Factory contract address */
  factoryAddress: string;
  /** Estimated gas cost in TON */
  estimatedGas: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ListJobsParams = {
  /**
   * Filter by job state
   */
  state?: ListJobsState;
  /**
   * Job type (TON or USDT)
   */
  type?: ListJobsType;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * @minimum 0
   */
  offset?: number;
};

export type ListJobsState = (typeof ListJobsState)[keyof typeof ListJobsState];

export const ListJobsState = {
  OPEN: "OPEN",
  FUNDED: "FUNDED",
  SUBMITTED: "SUBMITTED",
  COMPLETED: "COMPLETED",
  DISPUTED: "DISPUTED",
  CANCELLED: "CANCELLED",
} as const;

export type ListJobsType = (typeof ListJobsType)[keyof typeof ListJobsType];

export const ListJobsType = {
  TON: "TON",
  USDT: "USDT",
} as const;