# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ENACT Protocol Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * List ENACT jobs from the factory contract with optional filters
 * @summary List jobs
 */
export const listJobsQueryTypeDefault = `TON`;
export const listJobsQueryLimitDefault = 20;
export const listJobsQueryLimitMax = 100;

export const listJobsQueryOffsetDefault = 0;
export const listJobsQueryOffsetMin = 0;

export const ListJobsQueryParams = zod.object({
  state: zod
    .enum(["OPEN", "FUNDED", "SUBMITTED", "COMPLETED", "DISPUTED", "CANCELLED"])
    .optional()
    .describe("Filter by job state"),
  type: zod
    .enum(["TON", "USDT"])
    .default(listJobsQueryTypeDefault)
    .describe("Job type (TON or USDT)"),
  limit: zod.coerce
    .number()
    .min(1)
    .max(listJobsQueryLimitMax)
    .default(listJobsQueryLimitDefault),
  offset: zod.coerce
    .number()
    .min(listJobsQueryOffsetMin)
    .default(listJobsQueryOffsetDefault),
});

export const ListJobsResponse = zod.object({
  jobs: zod.array(
    zod.object({
      address: zod.string().describe("TON contract address"),
      jobId: zod.number().describe("Sequential job ID from factory"),
      state: zod
        .enum([
          "OPEN",
          "FUNDED",
          "SUBMITTED",
          "COMPLETED",
          "DISPUTED",
          "CANCELLED",
        ])
        .describe("Current state of the ENACT job"),
      type: zod.enum(["TON", "USDT"]),
      client: zod.string().describe("Client wallet address"),
      provider: zod
        .string()
        .nullish()
        .describe("Provider wallet address (null if not taken)"),
      evaluator: zod.string().describe("Evaluator wallet address"),
      budgetTon: zod
        .string()
        .describe("Budget in TON (string to avoid float precision issues)"),
      budgetUsdt: zod
        .string()
        .nullish()
        .describe("Budget in USDT (for Jetton jobs)"),
      descriptionHash: zod
        .string()
        .describe("IPFS\/TON Storage hash of job description"),
      descriptionText: zod
        .string()
        .nullish()
        .describe("Human-readable description (if available)"),
      resultHash: zod
        .string()
        .nullish()
        .describe("IPFS\/TON Storage hash of submitted result"),
      resultType: zod.number().nullish().describe("Result type code"),
      timeout: zod.number().describe("Job timeout in seconds"),
      evalTimeout: zod.number().describe("Evaluation timeout in seconds"),
      createdAt: zod.number().describe("Unix timestamp of creation"),
      submittedAt: zod
        .number()
        .nullish()
        .describe("Unix timestamp when result was submitted"),
      tonviewerUrl: zod
        .string()
        .describe("Link to TON explorer for this contract"),
    }),
  ),
  total: zod.number(),
  offset: zod.number(),
  limit: zod.number(),
});

/**
 * Get full details for a specific ENACT job by contract address
 * @summary Get job details
 */
export const GetJobParams = zod.object({
  address: zod.coerce.string().describe("TON contract address of the job"),
});

export const GetJobResponse = zod.object({
  address: zod.string().describe("TON contract address"),
  jobId: zod.number().describe("Sequential job ID from factory"),
  state: zod
    .enum(["OPEN", "FUNDED", "SUBMITTED", "COMPLETED", "DISPUTED", "CANCELLED"])
    .describe("Current state of the ENACT job"),
  type: zod.enum(["TON", "USDT"]),
  client: zod.string().describe("Client wallet address"),
  provider: zod
    .string()
    .nullish()
    .describe("Provider wallet address (null if not taken)"),
  evaluator: zod.string().describe("Evaluator wallet address"),
  budgetTon: zod
    .string()
    .describe("Budget in TON (string to avoid float precision issues)"),
  budgetUsdt: zod
    .string()
    .nullish()
    .describe("Budget in USDT (for Jetton jobs)"),
  descriptionHash: zod
    .string()
    .describe("IPFS\/TON Storage hash of job description"),
  descriptionText: zod
    .string()
    .nullish()
    .describe("Human-readable description (if available)"),
  resultHash: zod
    .string()
    .nullish()
    .describe("IPFS\/TON Storage hash of submitted result"),
  resultType: zod.number().nullish().describe("Result type code"),
  timeout: zod.number().describe("Job timeout in seconds"),
  evalTimeout: zod.number().describe("Evaluation timeout in seconds"),
  createdAt: zod.number().describe("Unix timestamp of creation"),
  submittedAt: zod
    .number()
    .nullish()
    .describe("Unix timestamp when result was submitted"),
  tonviewerUrl: zod.string().describe("Link to TON explorer for this contract"),
});

/**
 * Returns aggregate statistics for the ENACT protocol
 * @summary Get protocol statistics
 */
export const GetStatsResponse = zod.object({
  totalJobs: zod.number(),
  totalTonJobs: zod.number(),
  totalUsdtJobs: zod.number(),
  byState: zod.object({
    OPEN: zod.number(),
    FUNDED: zod.number(),
    SUBMITTED: zod.number(),
    COMPLETED: zod.number(),
    DISPUTED: zod.number(),
    CANCELLED: zod.number(),
  }),
  totalVolumeTon: zod.string().describe("Total volume in TON"),
  completionRate: zod.number().describe("Percentage of completed jobs (0-100)"),
  averageEvalTimeHours: zod
    .number()
    .nullish()
    .describe("Average evaluation time in hours"),
});

/**
 * Generate a Tonkeeper deeplink or transaction BOC for creating a new job
 * @summary Generate job creation link
 */
export const createJobLinkBodyTimeoutHoursDefault = 24;
export const createJobLinkBodyEvalTimeoutHoursDefault = 24;

export const CreateJobLinkBody = zod.object({
  evaluator: zod.string().describe("Evaluator wallet address"),
  budgetTon: zod.string().describe("Budget in TON"),
  description: zod.string().describe("Job description text"),
  timeoutHours: zod
    .number()
    .default(createJobLinkBodyTimeoutHoursDefault)
    .describe("Job timeout in hours"),
  evalTimeoutHours: zod
    .number()
    .default(createJobLinkBodyEvalTimeoutHoursDefault)
    .describe("Evaluation timeout in hours"),
  jobType: zod.enum(["TON", "USDT"]).optional(),
});

export const CreateJobLinkResponse = zod.object({
  tonkeeperLink: zod
    .string()
    .describe("Tonkeeper deeplink for transaction signing"),
  factoryAddress: zod.string().describe("Factory contract address"),
  estimatedGas: zod.string().describe("Estimated gas cost in TON"),
});