# lib/api-zod/src/generated/types/createJobLinkResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ENACT Protocol Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateJobLinkResponse {
  /** Tonkeeper deeplink for transaction signing */
  tonkeeperLink: string;
  /** Factory contract address */
  factoryAddress: string;
  /** Estimated gas cost in TON */
  estimatedGas: string;
}