# lib/api-zod/src/generated/types/createJobRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ENACT Protocol Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { JobType } from "./jobType";

export interface CreateJobRequest {
  /** Evaluator wallet address */
  evaluator: string;
  /** Budget in TON */
  budgetTon: string;
  /** Job description text */
  description: string;
  /** Job timeout in hours */
  timeoutHours?: number;
  /** Evaluation timeout in hours */
  evalTimeoutHours?: number;
  jobType?: JobType;
}