# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ENACT Protocol Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { JobState } from "./jobState";
import type { JobType } from "./jobType";

export interface Job {
  /** TON contract address */
  address: string;
  /** Sequential job ID from factory */
  jobId: number;
  state: JobState;
  type: JobType;
  /** Client wallet address */
  client: string;
  /** Provider wallet address (null if not taken) */
  provider?: string | null;
  /** Evaluator wallet address */
  evaluator: string;
  /** Budget in TON (string to avoid float precision issues) */
  budgetTon: string;
  /** Budget in USDT (for Jetton jobs) */
  budgetUsdt?: string | null;
  /** IPFS/TON Storage hash of job description */
  descriptionHash: string;
  /** Human-readable description (if available) */
  descriptionText?: string | null;
  /** IPFS/TON Storage hash of submitted result */
  resultHash?: string | null;
  /** Result type code */
  resultType?: number | null;
  /** Job timeout in seconds */
  timeout: number;
  /** Evaluation timeout in seconds */
  evalTimeout: number;
  /** Unix timestamp of creation */
  createdAt: number;
  /** Unix timestamp when result was submitted */
  submittedAt?: number | null;
  /** Link to TON explorer for this contract */
  tonviewerUrl: string;
}