# lib/api-zod/src/generated/types/jobListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ENACT Protocol Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from "./job";

export interface JobListResponse {
  jobs: Job[];
  total: number;
  offset: number;
  limit: number;
}