# lib/api-zod/src/generated/types/jobState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ENACT Protocol Dashboard API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Current state of the ENACT job
 */
export type JobState = (typeof JobState)[keyof typeof JobState];

export const JobState = {
  OPEN: "OPEN",
  FUNDED: "FUNDED",
  SUBMITTED: "SUBMITTED",
  COMPLETED: "COMPLETED",
  DISPUTED: "DISPUTED",
  CANCELLED: "CANCELLED",
} as const;