# lib/api-zod/src/generated/types/jobType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ENACT Protocol Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export type JobType = (typeof JobType)[keyof typeof JobType];

export const JobType = {
  TON: "TON",
  USDT: "USDT",
} as const;