# lib/api-zod/src/generated/types/protocolStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ENACT Protocol Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ProtocolStatsByState } from "./protocolStatsByState";

export interface ProtocolStats {
  totalJobs: number;
  totalTonJobs: number;
  totalUsdtJobs: number;
  byState: ProtocolStatsByState;
  /** Total volume in TON */
  totalVolumeTon: string;
  /** Percentage of completed jobs (0-100) */
  completionRate: number;
  /** Average evaluation time in hours */
  averageEvalTimeHours?: number | null;
}